fix money converting

+++ src/components/transactions/components/TransactionAddForm.tsx
@@ -49,7 +49,7 @@ export const TransactionAddForm: FC<TransactionAddFormProps> = ({ onTransactionA
         date: transactionDate,
         card: transactionCard,
         category: transactionCategoryID,
-        money: parseInt(transactionMoney?.current?.value as string),
+        money: parseFloat(transactionMoney?.current?.value as string),
         type: categoryType,
         notes: transactionNotes?.current?.value,
       },
